refactor hook to shared managementConsole hooks
